Convert image decoder helper class into a OnHeaderDecoded listener.
Avoiding inheritance makes it easier to reuse in the future. I also think it's a less confusing code structure. This is technically a breaking change, but the API is experimental anyway...

PiperOrigin-RevId: 422642039
