+++ library/src/main/java/com/bumptech/glide/load/resource/bitmap/BitmapImageDecoderResourceDecoder.java
@@ -2,30 +2,35 @@ package com.bumptech.glide.load.resource.bitmap;
 
 import android.graphics.Bitmap;
 import android.graphics.ImageDecoder;
-import android.graphics.ImageDecoder.OnHeaderDecodedListener;
 import android.graphics.ImageDecoder.Source;
 import android.util.Log;
+import androidx.annotation.NonNull;
 import androidx.annotation.RequiresApi;
+import com.bumptech.glide.load.Options;
+import com.bumptech.glide.load.ResourceDecoder;
 import com.bumptech.glide.load.engine.Resource;
 import com.bumptech.glide.load.engine.bitmap_recycle.BitmapPool;
 import com.bumptech.glide.load.engine.bitmap_recycle.BitmapPoolAdapter;
-import com.bumptech.glide.load.resource.ImageDecoderResourceDecoder;
+import com.bumptech.glide.load.resource.DefaultOnHeaderDecodedListener;
 import java.io.IOException;
 
-/** {@link Bitmap} specific implementation of {@link ImageDecoderResourceDecoder}. */
+/** {@link Bitmap} specific implementation of {@link DefaultOnHeaderDecodedListener}. */
 @RequiresApi(api = 28)
-public final class BitmapImageDecoderResourceDecoder extends ImageDecoderResourceDecoder<Bitmap> {
+public final class BitmapImageDecoderResourceDecoder implements ResourceDecoder<Source, Bitmap> {
   private static final String TAG = "BitmapImageDecoder";
   private final BitmapPool bitmapPool = new BitmapPoolAdapter();
 
   @Override
-  protected Resource<Bitmap> decode(
-      Source source,
-      int requestedResourceWidth,
-      int requestedResourceHeight,
-      OnHeaderDecodedListener listener)
-      throws IOException {
-    Bitmap result = ImageDecoder.decodeBitmap(source, listener);
+  public boolean handles(@NonNull Source source, @NonNull Options options) throws IOException {
+    return true;
+  }
+
+  @Override
+  public Resource<Bitmap> decode(
+      @NonNull Source source, int width, int height, @NonNull Options options) throws IOException {
+    Bitmap result =
+        ImageDecoder.decodeBitmap(
+            source, new DefaultOnHeaderDecodedListener(width, height, options));
     if (Log.isLoggable(TAG, Log.VERBOSE)) {
       Log.v(
           TAG,
@@ -36,9 +41,9 @@ public final class BitmapImageDecoderResourceDecoder extends ImageDecoderResourc
               + result.getHeight()
               + "]"
               + " for ["
-              + requestedResourceWidth
+              + width
               + "x"
-              + requestedResourceHeight
+              + height
               + "]");
     }
     return new BitmapResource(result, bitmapPool);
